Moved VideoCellOverlay to separate file
