+++ react/gameday2/gameday2.js
@@ -7,6 +7,7 @@ import GamedayNavbar from './components/GamedayNavbar';
 import EmbedTwitch from './components/EmbedTwitch';
 import Embedyoutube from './components/EmbedYoutube';
 import EmbedUstream from './components/EmbedUstream';
+import VideoCellOverlay from './components/VideoCellOverlay';
 
 var GamedayFrame = React.createClass({
   getInitialState: function() {
@@ -303,30 +304,6 @@ var VideoCell = React.createClass({
       )
     } else {
       return <div className={classes} />
-    }
-  }
-});
-
-var VideoCellOverlay = React.createClass({
-  onCloseClicked: function() {
-    this.props.onWebcastRemove(this.props.webcast);
-  },
-  render: function() {
-    var classes = classNames({
-      'hidden': !this.props.enabled,
-      'panel': true,
-      'panel-default': true,
-      'video-cell-overlay': true,
-    });
-    if (this.props.webcast) {
-      return (
-        <div className={classes}>
-          <div className="panel-heading">
-            <h3 className="panel-title">{this.props.webcast.name}</h3>
-            <span className="button-close glyphicon glyphicon-remove" onClick={this.onCloseClicked}></span>
-          </div>
-        </div>
-      )
     }
   }
 });
